requirements.txt
torch
torchvision
pyyaml
numpy
pandas
matplotlib
tqdm
scikit-learn